+++ OzzTradeDiary.NET/OzzTradeDiary/Models/Trade.part.cs
@@ -1,0 +1,9 @@
+﻿namespace TD.Models
+{
+    public partial class Trade
+    {
+        public decimal? ExecutedPositionValue => ExecutedEntry * FilledQuantity;
+        public decimal? PlannedPositionValue => PlannedEntry * OrderQuantity;
+
+    }
+}